Create task package. Refactors Task related classes into a package.
