+++ src/main/java/Duke.java
@@ -35,6 +35,6 @@ public class Duke {
     }
 
     public static void main(String[] args) {
-        new Duke("data/tasks.txt").run();
+        new Duke("data/duke.txt").run();
     }
 }
